style: Move AbsoluteColor to common color module
Differential Revision: https://phabricator.services.mozilla.com/D169607
