+++ components/style/values/specified/color.rs
@@ -5,7 +5,7 @@
 //! Specified color values.
 
 use super::AllowQuirks;
-use crate::color::ColorComponents;
+use crate::color::{AbsoluteColor, ColorComponents, ColorSpace};
 use crate::media_queries::Device;
 use crate::parser::{Parse, ParserContext};
 use crate::values::computed::{Color as ComputedColor, Context, ToComputedValue};
@@ -107,101 +107,7 @@ impl ColorMix {
     }
 }
 
-/// A color space representation in the CSS specification.
-///
-/// https://w3c.github.io/csswg-drafts/css-color-4/#color-type
-#[derive(Clone, Copy, Debug, MallocSizeOf, PartialEq, ToShmem)]
-#[repr(u8)]
-pub enum ColorSpace {
-    /// A color specified in the Lab color format, e.g.
-    /// "lab(29.2345% 39.3825 20.0664)".
-    /// https://w3c.github.io/csswg-drafts/css-color-4/#lab-colors
-    Lab,
-    /// A color specified in the Lch color format, e.g.
-    /// "lch(29.2345% 44.2 27)".
-    /// https://w3c.github.io/csswg-drafts/css-color-4/#lch-colors
-    Lch,
-    /// A color specified in the Oklab color format, e.g.
-    /// "oklab(40.101% 0.1147 0.0453)".
-    /// https://w3c.github.io/csswg-drafts/css-color-4/#lab-colors
-    Oklab,
-    /// A color specified in the Oklch color format, e.g.
-    /// "oklch(40.101% 0.12332 21.555)".
-    /// https://w3c.github.io/csswg-drafts/css-color-4/#lch-colors
-    Oklch,
-    /// A color specified with the color(..) function and the "srgb" color
-    /// space, e.g. "color(srgb 0.691 0.139 0.259)".
-    Srgb,
-    /// A color specified with the color(..) function and the "srgb-linear"
-    /// color space, e.g. "color(srgb-linear 0.435 0.017 0.055)".
-    SrgbLinear,
-    /// A color specified with the color(..) function and the "display-p3"
-    /// color space, e.g. "color(display-p3 0.84 0.19 0.72)".
-    DisplayP3,
-    /// A color specified with the color(..) function and the "a98-rgb" color
-    /// space, e.g. "color(a98-rgb 0.44091 0.49971 0.37408)".
-    A98Rgb,
-    /// A color specified with the color(..) function and the "prophoto-rgb"
-    /// color space, e.g. "color(prophoto-rgb 0.36589 0.41717 0.31333)".
-    ProphotoRgb,
-    /// A color specified with the color(..) function and the "rec2020" color
-    /// space, e.g. "color(rec2020 0.42210 0.47580 0.35605)".
-    Rec2020,
-    /// A color specified with the color(..) function and the "xyz-d50" color
-    /// space, e.g. "color(xyz-d50 0.2005 0.14089 0.4472)".
-    XyzD50,
-    /// A color specified with the color(..) function and the "xyz-d65" or "xyz"
-    /// color space, e.g. "color(xyz-d65 0.21661 0.14602 0.59452)".
-    XyzD65,
-}
-
-bitflags! {
-    #[derive(Default, MallocSizeOf, ToShmem)]
-    #[repr(C)]
-    struct SerializationFlags : u8 {
-        const AS_COLOR_FUNCTION = 0x01;
-    }
-}
-
-/// An absolutely specified color, using either rgb(), rgba(), lab(), lch(),
-/// oklab(), oklch() or color().
-#[derive(Clone, Debug, MallocSizeOf, PartialEq, ToShmem)]
-#[repr(C)]
-pub struct AbsoluteColor {
-    components: ColorComponents,
-    alpha: f32,
-    color_space: ColorSpace,
-    flags: SerializationFlags,
-}
-
-macro_rules! color_components_as {
-    ($c:expr, $t:ty) => {{
-        // This macro is not an inline function, because we can't use the
-        // generic  type ($t) in a constant expression as per:
-        // https://github.com/rust-lang/rust/issues/76560
-        const_assert_eq!(std::mem::size_of::<$t>(), std::mem::size_of::<[f32; 4]>());
-        const_assert_eq!(std::mem::align_of::<$t>(), std::mem::align_of::<[f32; 4]>());
-        const_assert!(std::mem::size_of::<AbsoluteColor>() >= std::mem::size_of::<$t>());
-        const_assert_eq!(
-            std::mem::align_of::<AbsoluteColor>(),
-            std::mem::align_of::<$t>()
-        );
-
-        std::mem::transmute::<&ColorComponents, &$t>(&$c.components)
-    }};
-}
-
 impl AbsoluteColor {
-    /// Create a new [AbsoluteColor] with the given [ColorSpace] and components.
-    pub fn new(color_space: ColorSpace, components: ColorComponents, alpha: f32) -> Self {
-        Self {
-            components,
-            alpha,
-            color_space,
-            flags: SerializationFlags::empty(),
-        }
-    }
-
     /// Convenience function to create a color in the sRGB color space.
     pub fn from_rgba(rgba: RGBA) -> Self {
         let red = rgba.red as f32 / 255.0;
@@ -215,12 +121,6 @@ impl AbsoluteColor {
         )
     }
 
-    /// Return the alpha component.
-    #[inline]
-    pub fn alpha(&self) -> f32 {
-        self.alpha
-    }
-
     /// Convert the color to sRGB color space and return it in the RGBA struct.
     pub fn to_rgba(&self) -> RGBA {
         let rgba = self.to_color_space(ColorSpace::Srgb);
@@ -230,178 +130,6 @@ impl AbsoluteColor {
         let blue = (rgba.components.2 * 255.0).round() as u8;
 
         RGBA::new(red, green, blue, rgba.alpha)
-    }
-
-    /// Convert this color to the specified color space.
-    pub fn to_color_space(&self, color_space: ColorSpace) -> Self {
-        use crate::color::convert;
-        use ColorSpace::*;
-
-        if self.color_space == color_space {
-            return self.clone();
-        }
-
-        let (xyz, white_point) = match self.color_space {
-            Lab => convert::to_xyz::<convert::Lab>(&self.components),
-            Lch => convert::to_xyz::<convert::Lch>(&self.components),
-            Oklab => convert::to_xyz::<convert::Oklab>(&self.components),
-            Oklch => convert::to_xyz::<convert::Oklch>(&self.components),
-            Srgb => convert::to_xyz::<convert::Srgb>(&self.components),
-            SrgbLinear => convert::to_xyz::<convert::SrgbLinear>(&self.components),
-            DisplayP3 => convert::to_xyz::<convert::DisplayP3>(&self.components),
-            A98Rgb => convert::to_xyz::<convert::A98Rgb>(&self.components),
-            ProphotoRgb => convert::to_xyz::<convert::ProphotoRgb>(&self.components),
-            Rec2020 => convert::to_xyz::<convert::Rec2020>(&self.components),
-            XyzD50 => convert::to_xyz::<convert::XyzD50>(&self.components),
-            XyzD65 => convert::to_xyz::<convert::XyzD65>(&self.components),
-        };
-
-        let result = match color_space {
-            Lab => convert::from_xyz::<convert::Lab>(&xyz, white_point),
-            Lch => convert::from_xyz::<convert::Lch>(&xyz, white_point),
-            Oklab => convert::from_xyz::<convert::Oklab>(&xyz, white_point),
-            Oklch => convert::from_xyz::<convert::Oklch>(&xyz, white_point),
-            Srgb => convert::from_xyz::<convert::Srgb>(&xyz, white_point),
-            SrgbLinear => convert::from_xyz::<convert::SrgbLinear>(&xyz, white_point),
-            DisplayP3 => convert::from_xyz::<convert::DisplayP3>(&xyz, white_point),
-            A98Rgb => convert::from_xyz::<convert::A98Rgb>(&xyz, white_point),
-            ProphotoRgb => convert::from_xyz::<convert::ProphotoRgb>(&xyz, white_point),
-            Rec2020 => convert::from_xyz::<convert::Rec2020>(&xyz, white_point),
-            XyzD50 => convert::from_xyz::<convert::XyzD50>(&xyz, white_point),
-            XyzD65 => convert::from_xyz::<convert::XyzD65>(&xyz, white_point),
-        };
-
-        Self::new(color_space, result, self.alpha)
-    }
-}
-
-impl From<cssparser::PredefinedColorSpace> for ColorSpace {
-    fn from(value: cssparser::PredefinedColorSpace) -> Self {
-        match value {
-            cssparser::PredefinedColorSpace::Srgb => ColorSpace::Srgb,
-            cssparser::PredefinedColorSpace::SrgbLinear => ColorSpace::SrgbLinear,
-            cssparser::PredefinedColorSpace::DisplayP3 => ColorSpace::DisplayP3,
-            cssparser::PredefinedColorSpace::A98Rgb => ColorSpace::A98Rgb,
-            cssparser::PredefinedColorSpace::ProphotoRgb => ColorSpace::ProphotoRgb,
-            cssparser::PredefinedColorSpace::Rec2020 => ColorSpace::Rec2020,
-            cssparser::PredefinedColorSpace::XyzD50 => ColorSpace::XyzD50,
-            cssparser::PredefinedColorSpace::XyzD65 => ColorSpace::XyzD65,
-        }
-    }
-}
-
-impl From<cssparser::AbsoluteColor> for AbsoluteColor {
-    fn from(f: cssparser::AbsoluteColor) -> Self {
-        match f {
-            cssparser::AbsoluteColor::Rgba(rgba) => Self::from_rgba(rgba),
-
-            cssparser::AbsoluteColor::Lab(lab) => Self::new(
-                ColorSpace::Lab,
-                ColorComponents(lab.lightness, lab.a, lab.b),
-                lab.alpha,
-            ),
-
-            cssparser::AbsoluteColor::Lch(lch) => Self::new(
-                ColorSpace::Lch,
-                ColorComponents(lch.lightness, lch.chroma, lch.hue),
-                lch.alpha,
-            ),
-
-            cssparser::AbsoluteColor::Oklab(oklab) => Self::new(
-                ColorSpace::Oklab,
-                ColorComponents(oklab.lightness, oklab.a, oklab.b),
-                oklab.alpha,
-            ),
-
-            cssparser::AbsoluteColor::Oklch(oklch) => Self::new(
-                ColorSpace::Oklch,
-                ColorComponents(oklch.lightness, oklch.chroma, oklch.hue),
-                oklch.alpha,
-            ),
-
-            cssparser::AbsoluteColor::ColorFunction(c) => {
-                let mut result = AbsoluteColor::new(
-                    c.color_space.into(),
-                    ColorComponents(c.c1, c.c2, c.c3),
-                    c.alpha,
-                );
-
-                if matches!(c.color_space, cssparser::PredefinedColorSpace::Srgb) {
-                    result.flags |= SerializationFlags::AS_COLOR_FUNCTION;
-                }
-
-                result
-            },
-        }
-    }
-}
-
-impl ToCss for AbsoluteColor {
-    fn to_css<W>(&self, dest: &mut CssWriter<W>) -> fmt::Result
-    where
-        W: Write,
-    {
-        match self.color_space {
-            ColorSpace::Srgb if !self.flags.contains(SerializationFlags::AS_COLOR_FUNCTION) => {
-                cssparser::ToCss::to_css(
-                    &cssparser::RGBA::from_floats(
-                        self.components.0,
-                        self.components.1,
-                        self.components.2,
-                        self.alpha(),
-                    ),
-                    dest,
-                )
-            },
-            ColorSpace::Lab => cssparser::ToCss::to_css(
-                unsafe { color_components_as!(self, cssparser::Lab) },
-                dest,
-            ),
-            ColorSpace::Lch => cssparser::ToCss::to_css(
-                unsafe { color_components_as!(self, cssparser::Lch) },
-                dest,
-            ),
-            ColorSpace::Oklab => cssparser::ToCss::to_css(
-                unsafe { color_components_as!(self, cssparser::Oklab) },
-                dest,
-            ),
-            ColorSpace::Oklch => cssparser::ToCss::to_css(
-                unsafe { color_components_as!(self, cssparser::Oklch) },
-                dest,
-            ),
-            _ => {
-                let color_space = match self.color_space {
-                    ColorSpace::Lab | ColorSpace::Lch | ColorSpace::Oklab | ColorSpace::Oklch => {
-                        unreachable!("Handle these in the wrapping match case!!")
-                    },
-                    ColorSpace::Srgb => {
-                        debug_assert!(
-                            self.flags.contains(SerializationFlags::AS_COLOR_FUNCTION),
-                             "The case without this flag should be handled in the wrapping match case!!"
-                          );
-
-                        cssparser::PredefinedColorSpace::Srgb
-                    },
-                    ColorSpace::SrgbLinear => cssparser::PredefinedColorSpace::SrgbLinear,
-                    ColorSpace::DisplayP3 => cssparser::PredefinedColorSpace::DisplayP3,
-                    ColorSpace::A98Rgb => cssparser::PredefinedColorSpace::A98Rgb,
-                    ColorSpace::ProphotoRgb => cssparser::PredefinedColorSpace::ProphotoRgb,
-                    ColorSpace::Rec2020 => cssparser::PredefinedColorSpace::Rec2020,
-                    ColorSpace::XyzD50 => cssparser::PredefinedColorSpace::XyzD50,
-                    ColorSpace::XyzD65 => cssparser::PredefinedColorSpace::XyzD65,
-                };
-
-                let color_function = cssparser::ColorFunction {
-                    color_space,
-                    c1: self.components.0,
-                    c2: self.components.1,
-                    c3: self.components.2,
-                    alpha: self.alpha,
-                };
-                let color = cssparser::AbsoluteColor::ColorFunction(color_function);
-                cssparser::ToCss::to_css(&color, dest)
-            },
-        }
     }
 }
